requirements.txt
aiofiles>=23.0.0
aiohttp>=3.8.0
PyYAML>=6.0
pathlib2>=2.3.7; python_version < "3.4"
rich>=13.0.0